Database のプロパティ、および、処理の一部を Namespace に移動
Database のプロパティの一部が Namespace 毎に個別に持つべき情報だった
ため、それらを関連する処理と共に Namespace に移動した。
